fix(datapoint): promote DataPoint.Value from float32 to float64
float32's 24-bit mantissa stair-steps on any counter above 2^24 (~16.7M),
causing 64-bit SNMP ifInOctets/ifOutOctets and large fs-byte gauges to
round to the nearest 16-byte step. float64's 53-bit mantissa preserves
those values exactly.

Migration:
- DataPoint.Value float64 (types/datapoint/types.go)
- SenhubDataPoint.Value float64 (senhub strategy — we own both ends)
- PRTG wire format keeps float32 at the boundary (explicit cast); PRTG
  spec uses float JSON, which doesn't distinguish 32 vs 64, and the
  cast is intentional and documented
- All probe emit sites updated (explicit casts removed now that the
  field is float64)
- All test fixtures and golden outputs regenerated; the Nagios golden
  files now show real metric values instead of empty status=0 lines
  (the float32 precision gap was silently dropping data in the mapper)

Closes #271

+++ internal/agent/services/data_store/data_store.go
@@ -543,10 +543,9 @@ func (d *dataStore) applyUnitCorrections(datapoints []datapoint.DataPoint) []dat
 		if defTransformer, ok := transformer.(interface {
 			ApplyUnitCorrection(string, float64, map[string]string) (float64, bool)
 		}); ok {
-			// Convert value to float64 for correction calculation
-			originalFloat64 := float64(dp.Value)
+			originalFloat64 := dp.Value
 			if newValue, applied := defTransformer.ApplyUnitCorrection(dp.Name, originalFloat64, tagMap); applied {
-				correctedValue = float32(newValue)
+				correctedValue = newValue
 				correctionCount++
 
 				d.logger.Info().
